requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
gradio>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
numpy>=1.24.0

